add views folder and each view in a unique .py file
