+++ py_backend/users/views/__init__.py
@@ -1,0 +1,8 @@
+from .register import *
+from .login import *
+from .logout_view import *
+from .update_profile_username import *
+from .update_profile_bio import *
+from .update_profile_password import *
+from .update_profile_picture import *
+from .get_csrf_token import *
